[fix] 팝업 굿즈 결제 기능 문제 해결
