+++ backend/src/main/java/com/fiiiiive/zippop/domain/orders/dto/RefundEvent.java
@@ -1,0 +1,11 @@
+package com.fiiiiive.zippop.domain.orders.dto;
+
+import com.siot.IamportRestClient.response.Payment;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class RefundEvent {
+    private Payment payment;
+}
