+++ backend/src/main/java/com/fiiiiive/zippop/domain/orders/entity/OrdersDetail.java
@@ -1,8 +1,8 @@
-package com.fiiiiive.zippop.orders.entity;
+package com.fiiiiive.zippop.domain.orders.entity;
 
 import com.fiiiiive.zippop.global.base.BaseEntity;
-import com.fiiiiive.zippop.goods.entity.Goods;
-import com.fiiiiive.zippop.orders.dto.OrdersDto;
+import com.fiiiiive.zippop.domain.goods.entity.Goods;
+import com.fiiiiive.zippop.domain.orders.dto.OrdersDto;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.*;
