+++ backend/src/main/java/com/fiiiiive/zippop/domain/orders/repository/OrdersDetailRepository.java
@@ -1,7 +1,7 @@
-package com.fiiiiive.zippop.orders.repository;
+package com.fiiiiive.zippop.domain.orders.repository;
 
 
-import com.fiiiiive.zippop.orders.entity.OrdersDetail;
+import com.fiiiiive.zippop.domain.orders.entity.OrdersDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
